FInished with the documentation.
There are some diffs: if src/d is a file/dir in the first quic and
 dir/file in the next remove dest/d and copy the proper element from src.
Also added a counter for the deleted elements.
Finally, coded the non -l command to ignore the symlink and treat the hardlinks as files. if we include -l then every link is treated properly.